% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.var.compare.R
\name{cor.var.compare}
\alias{cor.var.compare}
\title{Differential correlations between a single variable and all variables within a dataset}
\usage{
cor.var.compare(variable, data, group)
}
\description{
Differential correlations between a single variable and all variables within a dataset
}
\author{
Emily Mears, \email{mears.emilyrose@gmail.com}
}
