% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.cor.R
\name{data.cor}
\alias{data.cor}
\title{Differential correlations between all variables within a dataset}
\usage{
data.cor(data, group, ordered = "None", limit = NA)
}
\arguments{
\item{data}{the dataset for analysis}

\item{group}{the grouping variable for comparative differential correlations}

\item{ordered}{which column should the table be ordered by?}

\item{limit}{how many combinations should the table display}
}
\value{
a table (or dataframe) with the correlation coefficients, p-values, fisher r-to-z statistic and BH p-value correlation for each correlation pair
}
\description{
Differential correlations between all variables within a dataset
}
\author{
Emily Mears, \email{mears.emilyrose@gmail.com}
}
