% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.cor.vis.R
\name{gg.cor.vis}
\alias{gg.cor.vis}
\title{Differential correlation visualisation}
\usage{
gg.cor.vis(var_1, var_2, data, group,
  group_1_title = "Group 1 Correlation",
  group_2_title = "Group 2 Correlation")
}
\arguments{
\item{var_1}{character string indicating the first variable for differential correlation}

\item{var_2}{character string indicating the second variable for differential correlation}

\item{data}{dataframe for analysis}

\item{group}{character string defining the grouping variable for comparative differential correlations}

\item{group_1_title}{character string defining the first plot title}

\item{group_2_title}{character string defining the title second plot title}
}
\value{
dual correlation plot for each group with correlation coefficents and p-values
}
\description{
Creates two correlation plots between two variables, separated by a grouping variable containing two groups.
}
\examples{
## Load example dataframes
df <- read.csv("example_data/excorr_df1.csv")

## Run function
gg.cor.vis(var_1 = "NTRpFI", var_2 = "X70J8B6", df, group = "sex")
gg.cor.vis(var_1 = "NTRpFI", var_2 = "X70J8B6", df, group = "sex", group_1_title = "Group 1 Correlation", group_2_title = "Group 2 Correlation")

}
\author{
Emily Mears, \email{mears.emilyrose@gmail.com}, Matthew Grant

Ben Day, \email{benjamindayengineer@gmail.com}
}
