% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.cor.vis.R
\name{gg.corr.vis}
\alias{gg.corr.vis}
\title{Differential correlation visualisation}
\usage{
gg.corr.vis(var_1, var_2, data, group)
}
\description{
Creates two correlation plots between two variables, separated by a grouping variable
}
\author{
Emily Mears, \email{mears.emilyrose@gmail.com}
}
