% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.cor.var.compare.r
\name{multi.cor.var.compare}
\alias{multi.cor.var.compare}
\title{Calculates differential correlation statistics between a specified variable and all other variables within a different dataset.}
\usage{
multi.cor.var.compare(variable, data1, data2, sample_col, group,
  ordered = "fisher")
}
\arguments{
\item{variable}{character string indicating the variable of interest for differential correlation (within \code{data1})}

\item{data1}{first dataframe containing variable of interest}

\item{data2}{second dataframe containing variable to compare}

\item{sample_col}{character string defining the sample or identifying column in both datasets (common to both dataframes)}

\item{group}{character string defining the grouping variable for comparative differential correlations}

\item{ordered}{character string defining which column the table should be ordered by. Choose from \code{g1cor}, \code{g1p}, \code{g2cor}, \code{g2r}, \code{fisher} (default) and \code{BH}.}
}
\value{
a table (or dataframe) with Pearson correlation coefficients (r), associated p-values, fisher r-to-z statistic and BH p-value correlation for each correlation pair
}
\description{
Creates a table of pairwise correlations statistics, with separation of two groups for comparison.
This can be used as an exploratory tool to investigate correlations between a specific variable of interest and all other variables within a separate dataset.
}
\examples{
## Load example dataframes
df1 <- read.csv("example_data/excorr_df1.csv")
df2 <- read.csv("example_data/excorr_df2.csv")

## Run function
multi.cor.var.compare(variable = "NTRpFI", df1, df2, sample_col = "sample", group = "sex")
multi.cor.var.compare(variable = "NTRpFI", df1, df2, sample_col = "sample", group = "sex", ordered = "fisher")

}
\author{
Emily Mears, \email{mears.emilyrose@gmail.com}, Matthew Grant, \email{mgra576@aucklanduni.ac.nz}

Ben Day, \email{benjamindayengineer@gmail.com}
}
