% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.data.cor.r
\name{multi.data.cor}
\alias{multi.data.cor}
\title{Calculates differential correlation statistics for all the possible variable-variable combinations between two datasets.}
\usage{
multi.data.cor(data1, data2, sample_col, common_cols, group,
  ordered = "fisher", limit = NA)
}
\arguments{
\item{data1}{first dataframe}

\item{data2}{second dataframe}

\item{sample_col}{character string defining the sample or identifying column in both dataframes}

\item{group}{character string defining the grouping variable for comparative differential correlations}

\item{ordered}{character string defining the variable to order output by. Choose from \code{g1cor}, \code{g1p}, \code{g2cor}, \code{g2r}, \code{fisher} (default) and \code{BH}.}

\item{limit}{numeric input to limit number of output correlation pairings}
}
\value{
a table (or dataframe) with the correlation coefficients, p-values, fisher r-to-z statistic and BH p-value correlation for each correlation pair
}
\description{
Requires grouping variable to contain 2 groups. Requires same samples across datasets (arranged in same order).
Declare columns (descriptors) column to both datasets.
}
\examples{
## Load example dataframes
df1 <- read.csv("example_data/excorr_df1.csv")
df2 <- read.csv("example_data/excorr_df2.csv")

## Run function
multi.data.cor(df1, df2, sample_col = "sample", common_cols = c("sex", "sample"), group = "sex")
multi.data.cor(df1, df2, sample_col = "sample", common_cols = c("sex", "sample"), group = "sex", ordered = "Group_2_Pvalue", limit = 100)

}
\author{
Emily Mears, \email{mears.emilyrose@gmail.com}, Matthew Grant, \email{mgra576@aucklanduni.ac.nz}

Ben Day, \email{benjamindayengineer@gmail.com}
}
