% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.var.compare.R
\name{cor.var.compare}
\alias{cor.var.compare}
\title{Differential correlations between a single variable and all variables within a dataset}
\usage{
cor.var.compare(variable, data, group)
}
\arguments{
\item{variable}{the variable of interest for differential correlation}

\item{data}{the dataset for analysis}

\item{group}{the grouping variable for comparative differential correlations}
}
\value{
a table (or dataframe) with the correlation coefficients, p-values, fisher r-to-z statistic and BH p-value correlation for each correlation pair
}
\description{
Differential correlations between a single variable and all variables within a dataset
}
\author{
Emily Mears, \email{mears.emilyrose@gmail.com}
}
