% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.cor.R
\name{data.cor}
\alias{data.cor}
\title{Differential correlations between all variables within a dataset}
\usage{
data.cor(data, group, ordered = "None", limit = NA)
}
\description{
Differential correlations between all variables within a dataset
}
\author{
Emily Mears, \email{mears.emilyrose@gmail.com}
}
